+++ RigidBody.hpp
@@ -22,7 +22,7 @@ namespace spic {
     class RigidBody : public Component {
         public:
 
-            RigidBody(double mass, double gravityScale, BodyType bodyType);
+            RigidBody(float mass, float gravityScale, BodyType bodyType);
             /**
              * @brief Apply force to this rigid body.
              * @param forceDirection A point, used as a vector to indicate direction
@@ -30,13 +30,13 @@ namespace spic {
              */
             void AddForce(const Vector2<float>& forceDirection);
 
-            double GetMass();
+            float GetMass();
 
-            void SetMass(double newMass);
+            void SetMass(float newMass);
 
-            double GetGravityScale();
+            float GetGravityScale();
 
-            void SetGravityScale(double newScale);
+            void SetGravityScale(float newScale);
 
             BodyType GetBodyType();
